+++ packages/client/test/unit/OrderMessages.test.ts
@@ -19,9 +19,10 @@ import { OrderMessages } from '../../src/subscribe/ordering/OrderMessages'
 import { fromArray } from '../../src/utils/GeneratorUtils'
 import { PushPipeline } from '../../src/utils/PushPipeline'
 import { MOCK_CONTENT } from '../test-utils/utils'
+import { randomBytes } from 'crypto'
 
 const STREAM_PART_ID = StreamPartIDUtils.parse('stream#0')
-const PUBLISHER_ID = randomEthereumAddress()
+const PUBLISHER_ID = randomBytes(20)
 const MSG_CHAIN_ID = 'mock-msg-chain-id'
 
 const CONFIG = {