+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changes before Tatum release are not documented in this file.
 #### Removed
 
 - Removed obsolete RPC provider options (https://github.com/streamr-dev/network/pull/2583)
+- Remove support for legacy encryption keys
 
 #### Fixed